requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
python-dotenv==1.0.0
youtube-dl==2021.12.17
pydub==0.25.1
librosa==0.10.0
numpy==1.24.3
eyed3==0.9.7
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
requests==2.31.0